+++ packages/tsurlfilter/src/configuration.ts
@@ -1,3 +1,7 @@
+import { version } from '../package.json';
+
+export const TSURLFILTER_VERSION = version;
+
 /**
  * Compatibility types are used to configure engine for better support of different libraries
  * For example: