+++ packages/tswebextension/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2023-12-06
 
+### Added
+- Library version number to the exports [#2237](https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2237).
+
 ## Fixed
 - Remove referrer from the document.referrer [#1844](https://github.com/AdguardTeam/AdguardBrowserExtension/issues/1844)